+++ Ranger/UI/TournamentTableWindow/TournamentWindowController.swift
@@ -9,7 +9,7 @@
 import Cocoa
 
 
-class TourneyWindowController: NSWindowController
+class TournamentWindowController: NSWindowController
 {
     
     
@@ -18,19 +18,19 @@ class TourneyWindowController: NSWindowController
     
     // MARK: - Lifecycle
     
-    static func instantiateAndShow(forTableWindowInfo tableWindowInfo: TableWindowInfo) -> TourneyWindowController
+    static func instantiateAndShow(forTableWindowInfo tableWindowInfo: TableWindowInfo) -> TournamentWindowController
     {
         // Instantiate a new controller.
         let storyboard = NSStoryboard(name: "Main", bundle: nil)
-        let tourneyTableWindowController = storyboard.instantiateController(withIdentifier: "TourneyTableWindow") as! TourneyWindowController
-            tourneyTableWindowController.showWindow(self)
+        let tournamentTableWindowController = storyboard.instantiateController(withIdentifier: "TournamentWindowController") as! TournamentWindowController
+            tournamentTableWindowController.showWindow(self)
         
         // Track table.
-        let tourneyTableViewController = tourneyTableWindowController.contentViewController as! TourneyViewController
+        let tourneyTableViewController = tournamentTableWindowController.contentViewController as! TournamentViewController
             tourneyTableViewController.track(tableWindowInfo)
         
         // Inject reference.
-        tourneyTableWindowController.tableWindowInfo = tableWindowInfo
+        tournamentTableWindowController.tableWindowInfo = tableWindowInfo
         
         // Hide buttons.
         if let window = tourneyTableViewController.view.window
@@ -40,12 +40,12 @@ class TourneyWindowController: NSWindowController
             window.standardWindowButton(.closeButton)?.isHidden = true
         }
         
-        return tourneyTableWindowController
+        return tournamentTableWindowController
     }
     
     func update(with tableWindowInfo: TableWindowInfo)
     {
-        let tourneyTableViewController = self.contentViewController as! TourneyViewController
+        let tourneyTableViewController = self.contentViewController as! TournamentViewController
             tourneyTableViewController.update(with: tableWindowInfo)
     }
 }